PageUserList implementation - is now showing users from database and it is able to delete selected user from the database.

+++ fidm/src/main/java/com/esuta/fidm/model/ModelService.java
@@ -63,7 +63,7 @@ public class ModelService implements IModelService{
     }
 
     @Override
-    public <T extends ObjectType> Integer countObjects(Class<T> type) throws DatabaseCommunicationException {
+    public <T extends ObjectType> Long countObjects(Class<T> type) throws DatabaseCommunicationException {
         return repositoryService.countObjects(type);
     }
